Code refactoring and organization done, removed unused components.
