+++ src/app/components/Projects.tsx
@@ -1,0 +1,25 @@
+import React from "react";
+import { Typography, Grid, Box } from "@mui/material";
+
+import ProjectCard from "../layout/ProjectCard";
+import { useData } from "../context/DataContext";
+
+const Projects = () => {
+    const projects = useData().projects;
+  return (
+    <Box sx={{ my: 4 }}>
+      <Typography variant="h4" gutterBottom>
+        PROJECTS
+      </Typography>
+      <Grid container spacing={3}>
+        {projects.map((project, index) => (
+          <Grid item xs={12} sm={6} md={4} key={index}>
+            <ProjectCard project={project} />
+          </Grid>
+        ))}
+      </Grid>
+    </Box>
+  );
+};
+
+export default Projects;
