+++ src/app/layout/App.tsx
@@ -2,7 +2,6 @@ import "./index.css";
 import React, { useState } from "react";
 import CssBaseline from "@mui/material/CssBaseline";
 import { PaletteMode, createTheme, Container } from "@mui/material";
-// import { extendTheme } from '@mui/joy/styles';
 import { ThemeProvider } from "@emotion/react";
 import AnimatedSection from "../animations/AnimatedSection";
 import Header from "../components/Header";
@@ -12,16 +11,12 @@ import About from "../components/About";
 import Footer from "../components/Footer";
 import ResumeDownloadButton from "../components/ResumeDownloadButton";
 import NavigateHomeButton from "../components/NavigateHomeButton";
-import WorkExperience from "./WorkExComponent";
-import ProjectComponent from "./ProjectComponent";
-import { useData } from "../context/DataContext";
+import WorkExperience from "../components/WorkExperience";
+import Projects from "../components/Projects";
 function App() {
   const [darkMode, setDarkMode] = useState(true);
   const [showNavUpBtn, setShowNavUpBtn] = useState(false);
   const darkTheme: PaletteMode = darkMode ? "dark" : "light";
-  const experiences = useData().workex;
-
-  const projects = useData().projects;
   const theme = createTheme({
     palette: {
       mode: darkTheme,
@@ -45,20 +40,17 @@ function App() {
   const sections = [
     { component: <Hero />, id: "home" },
     { component: <About />, id: "about" },
-    // { component: <Skills />, id: "skills" },
     { component: <Education />, id: "education" },
-    { component: <WorkExperience experiences={experiences} />, id: "workex" },
-    { component: <ProjectComponent projects={projects} />, id: "projects" },
+    { component: <WorkExperience />, id: "workex" },
+    { component: <Projects />, id: "projects" },
   ];
 
   window.addEventListener("scroll", () => {
     if (window.scrollY === 0) setShowNavUpBtn(false);
     if (!showNavUpBtn && window.scrollY !== 0) setShowNavUpBtn(true);
   });
-
   return (
     <>
-      {/* <Background3D /> */}
       <ThemeProvider theme={theme}>
         <CssBaseline />
 
@@ -67,7 +59,6 @@ function App() {
           disableGutters
           sx={{
             maxWidth: "80%",
-            //minWidth: "500px",
           }}
         >
           {showNavUpBtn ? <NavigateHomeButton /> : null}
